Factors RunStderr in to it's own library.
