+++ lib/btrfs/btrfs.go
@@ -17,6 +17,7 @@ import (
 	"time"
 
 	"code.google.com/p/go-uuid/uuid"
+	"github.com/pachyderm/pfs/lib/shell"
 )
 
 var letters = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
@@ -33,25 +34,8 @@ func RandSeq(n int) string {
 	return string(b)
 }
 
-func RunStderr(c *exec.Cmd) error {
-	stderr, err := c.StderrPipe()
-	if err != nil {
-		return err
-	}
-	err = c.Start()
-	if err != nil {
-		return err
-	}
-	buf := new(bytes.Buffer)
-	buf.ReadFrom(stderr)
-	if buf.Len() != 0 {
-		log.Print("Command had output on stderr.\n Cmd: ", strings.Join(c.Args, " "), "\nstderr: ", buf)
-	}
-	return c.Wait()
-}
-
 func Sync() error {
-	return RunStderr(exec.Command("sync"))
+	return shell.RunStderr(exec.Command("sync"))
 }
 
 func BasePath(name string) string {
@@ -167,29 +151,29 @@ func LazyWalk(name string, f func(string) error) error {
 }
 
 func SubvolumeCreate(name string) error {
-	return RunStderr(exec.Command("btrfs", "subvolume", "create", FilePath(name)))
+	return shell.RunStderr(exec.Command("btrfs", "subvolume", "create", FilePath(name)))
 }
 
 func SubvolumeDelete(name string) error {
-	return RunStderr(exec.Command("btrfs", "subvolume", "delete", FilePath(name)))
+	return shell.RunStderr(exec.Command("btrfs", "subvolume", "delete", FilePath(name)))
 }
 
 func Snapshot(volume string, dest string, readonly bool) error {
 	if readonly {
-		return RunStderr(exec.Command("btrfs", "subvolume", "snapshot", "-r",
+		return shell.RunStderr(exec.Command("btrfs", "subvolume", "snapshot", "-r",
 			FilePath(volume), FilePath(dest)))
 	} else {
-		return RunStderr(exec.Command("btrfs", "subvolume", "snapshot",
+		return shell.RunStderr(exec.Command("btrfs", "subvolume", "snapshot",
 			FilePath(volume), FilePath(dest)))
 	}
 }
 
 func SetReadOnly(volume string) error {
-	return RunStderr(exec.Command("btrfs", "property", "set", FilePath(volume), "ro", "true"))
+	return shell.RunStderr(exec.Command("btrfs", "property", "set", FilePath(volume), "ro", "true"))
 }
 
 func UnsetReadOnly(volume string) error {
-	return RunStderr(exec.Command("btrfs", "property", "set", FilePath(volume), "ro", "false"))
+	return shell.RunStderr(exec.Command("btrfs", "property", "set", FilePath(volume), "ro", "false"))
 }
 
 func CallCont(c *exec.Cmd, cont func(io.ReadCloser) error) error {
